requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
Django==4.0.5
django-cors-headers==3.12.0
django-environ==0.8.1
django-filter==21.1
django-stubs==1.11.0
django-stubs-ext==0.4.0
djangorestframework==3.13.1
gunicorn==20.1.0
isort==5.10.1
mypy==0.950
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
types-pytz==2021.3.8
types-PyYAML==6.0.8
typing_extensions==4.2.0
